Add functions for FE_grid class and Cell class
"set" and "get" functions added to write and read the variables nodes,
triangulation, grid, polybasis, neighbors from the classes fe_grid and
Cell. The constructors need to be finished.

For the cell class a specific Square class was defined also inheriting
most of the functions from Cell. Implementation should be verified.

A way to read files with grids done with gmsh has to be implemented in
order to store them into a fe_grid object and try PDE problems.